[plugins] maint: Keep zIndex when user change region. #407.
* qncllayout/src/ui/view/qnlygraphicsregion.cpp: Likewise. zIndex is an
  integer now. This commit fixes the bug #407.
* qncllayout/src/ui/view/qnlygraphicsregion.h: Likewise.
* qncllayout/src/ui/view/qnlygraphicsregionbase.cpp: Likewise.

+++ composer-plugins/qncllayout/src/ui/view/qnlygraphicsregion.h
@@ -106,9 +106,9 @@ public:
 
     void setColor(const QString &color);
 
-    qreal getzIndex() const;
+    int getzIndex() const;
 
-    void setzIndex(qreal zindex);
+    void setzIndex(int zindex);
 
     qreal getzIndexTop() const;
